Add support for default NLA Track names and improve the component's UI.
Handle the glTF exporter replacing the default track names
(NlaTrack and [Action Stash]) with the action names, and make
sure this is nicely conveyed by the UI.

Improve the component UI by redrawing immediately on changes,
using the shape key icon for shape key tracks, and making the
user select a new track from a menu instead of typing the new
name in themselves (this is required to handle default track names).

Improve error reporting slightly, now tracks that contain multiple actions,
no actions, actions that have since been renamed, or have names
(track or action depending on whether the track name is a default) containing
forbidden characters, will display with error icons.

Note: Invalid tracks can still be added as long as they aren't already present.
Duplicate tracks will only be shown once in the list (as they will be combined
on export).  Duplicate tracks are determined based on track name (if it's not
a default name) or action name (if it is).  If you have duplicate object and
shape key tracks, only the object track will be shown in the add menu.

Note2: Any files saved on a previous beta are not migrated automatically;
users will need to update or re-add their tracks (the same as if the name
had changed).