Improve error reporting and behavior.
add a proper error reporting system.

add a menu id to allow selection of shape key tracks with default names that match object tracks with default names.

rework the is_unique_action function to only check within the current object instead of checking for duplicate strip actions in the entire blender file.

rework the has_track to allow updating an invalid track to an NLA track with the same custom name (needed for when you have a combined object/shape key track and you rename the object track, which invalidates the loop-animation track, and then you want to update the loop-animation track to the shape key track or vice versa).

rework the is_matching_track function to match the track type and to report that the action has changed.

rework the is_useable_nla_track function to only check for errors once, check for additional errors, and to improve clarity.