Fix #123
This fixes bone components not being imported if they aren't in the
top level of the hierarchy.

Note:
The gltf bones aren't in the same order as the armature bones, this
is by design.  I considered using the bone order to map the gltf
bones to the armature bones (potentially faster), but decided it would
be too fragile, and mapping it would have potentially been slower anyway.

I loop through the bones in the armature and access the gltf bones
by name rather than the other way around because Blender's internal
data structures are mostly based on linked lists, so looping over it
is much more performant than doing lookups for each item.