Add handler for append/link.
Redirect the output of sys.stdout for bpy.context.window_manager.print_undo_steps()
to allow for checking for append/link operations on a depsgraph update.
Modify the migration handler to support local migrations and check whether
the gizmo system needs to be activated.  Refactor migrations to be performed
on individual components so they're simpler and more performant.