Fix the append/link handler not running when other Blender windows are open.
It turns out that space_data is invalid when more than one Blender window
is open, so add a check for if the file has just loaded.  Only skip the
handler if a file has just been loaded and space_data is invalid
(i.e. it will crash if the handler is run in this state).