Optimize update_gizmos calls.
Add a task scheduler to the undo_stack_handler to allow performance heavy
tasks (currently only update_gizmos) to be combined and executed at the end
of the handler.
Add an option to not call update_gizmos directly in the migration function
and handle it outside.
Update the add/remove component functions to only update gizmos if they're
adding a gizmo component.
Force a depsgraph update at the end of the add/remove component operators
so that the undo_stack_handler will handle them immediately, rather than
delaying them to a future undo step which may be a suboptimal time for a
gizmo update.