Include the name of the blend file in reflection probes' baked image file names.
This prevents name collisions when using the default relative path for
reflection probe bakes and there are multiple blend files in the same folder.

If the blend file hasn't been saved yet, "untitled" is used as the blend file name.

Name conflicts are still possible if the user sets tmp_path to a specific folder
and bakes from multiple, identically named blend files with one or more identically
named reflection probes.  E.g. if tmp_path is set to a specific folder, the folder
name is used as the project name, "scene" is used as the blend file name in a bunch
of projects, and the default names for reflection probes are kept, this will still
cause name conflicts.