Transfer reflection probe bakes from untitled files to tmp_path directory on save.
Redirect reflection probe bakes from untitled files to Blender's session temp directory
to avoid any permission errors and uncertainty from relative paths.  Then transfer the
files to the directory specified in tmp_path on save.
Note: Saving a copy of the blend file won't trigger a transfer or rename and the copy
will continue to look for the files in Blender's session temp directory.
Note2: In order to update the paths in the blend file to the new image location the
blend file has to be saved again.  To do this without destroying Blender's backup
save files, the backup (save_version) preference is stored, set to zero
before saving, and then restored to its original value afterward.

Display a warning underneath the tmp_path field in the preferences to inform the user
of the redirection and transfer.