Pack generated reflection probe environment maps instead of outputting them to a directory.
Use Blender's tempdir to render the probe environment maps.
Load the rendered environment maps and remove any old environment maps in the process if the name matches to simulate an overwrite and keep the blend file as clean as possible.  Users of the overwritten environment maps are remapped to use the new environment maps.
Pack the environment maps and delete the tempdir files so they are stored solely within Blender and change the image datablock/packed file file paths to reflect this.
Remove the tmp_path preference from the add-on preferences as it is no longer being used.

Note: an error is generated when changing the file paths on the environment maps after packing, but everything seems to work correctly.