Add -old to conflicting environment maps to improve reflection probe baking stability.
This helps prevent accidental overwrites of other reflection probe environment maps
when baking, due to Blender's name juggling and also allows name juggled environment
maps to be more easily spotted and dealt with by the user.  However it's still
possible to overwrite another probe's baked environment map if you try hard enough
and it's also possible for the -old postfix to end up in the name multiple times
through subsequent bakes/name juggles.

Further stabilization can be obtained by adding -rp (or any separating characters)
to the end of the environment map name before attending to conflicting environment
maps, however this will prevent overwrites of environment maps created before the
naming scheme change, and so will need a migration which can't be added until after
the 1.0.0 release.  Without a migration other users of the environment maps wouldn't
get updated when the probe is rebaked the first time, assuming they would have
gotten updated had the naming scheme not been changed.
Note: Adding -rp should fix both issues with the -old method.

Ideally, probe name changes could be detected and the environment maps renamed to
match so that just checking the probe's name against the environment map's name
would be enough, however this doesn't seem to be possible/practical/reliable in
Blender at this time.

For posterity, a scenario in which bakes could be overwritten:
Name juggling occurs because of some other probe being renamed to a conflicting name
and being baked, causing Probe A's environment map to get renamed with a -old postfix.
Probe A is duplicated.  Probe B (the duplicate) is renamed to match the environment
map (e.g. Probe A-old). Probe B (now called Probe A-old) is then baked.

For a scenario where the -old postfix would end up in the name multiple times,
follow the above steps for overwriting bakes, but remove the environment map for
Probe B (then called Probe A-old) before baking it.