Fix reflection probes sometimes failing to render and getting stuck mid-bake.
When rendering with UI updates, Blender opens a separate rendering thread,
and while this thread is open no new renders can be initiated.  Waiting
for half a second usually allowed this thread to close and a new render
to be started, however occasionally the thread would still be open and
the new render would return cancelled which resulted in the reflection
probe getting stuck in mid bake.  In order to work around this, probes
are initially set up and then rendering is repeatedly attempted until
it works.  For more details see: https://developer.blender.org/T52258