Allow reflection probes to be locked.
Add a button to the reflection probe panel to lock/unlock probes.
The Bake, Open EnvMap buttons and selected image field are disabled
when the probe is locked.  Locked probes are skipped when baking
or importing (a warning is shown on import).  Locked probes can be
included or not when exporting environment maps. The scene level
reflection probe settings are displayed even if all probes are locked.