Account for linked reflection probes.
Skip linked probes for all bakes, exports, imports, and opens.
Include a mention of linked probes in warnings.
Add warnings for linked probes when baking or opening an EnvMap for the active probe.
Refactor how linked data-blocks are detected.
Don't remove images when overwriting on import if the image is linked.