Support linked PointerProperties in Blender 3.2+
Manually disable the PointerProperty field when it's part of a linked component.
Add a disabled link indicator button to PointerProperty fields that point to linked
data-blocks.  This indicates to the user that the data-block is either linked or
overridden (depending on the icon) and when hovered over will pop up a tooltip
displaying the data-block name as well as its source library.
Note: The link indicator button is needed for Blender versions before 3.2 as well
because Blender doesn't always show the source library in its default tooltips.

Created a generic tooltip label operator that functions similarly to a regular
label, but will open a custom tooltip when hovered.  This is used to provide
the link indicator button.  Note: if the text of the tooltip label is being
truncated in the UI then the first line of the tooltip will show the text
without truncation. If the tooltip label isn't being truncated, then the
first line will contain three dashes, provided the tooltip has custom text,
otherwise the first line will be blank.  If there is no custom text, the
tooltip will display a single period in the second line
(Blender automatically adds one by default).