Switch to component specific versions.
Switch from storing the add-on version on the component instances to using individual component versions in the definitions and on each instance.
Clearly differentiate the names in the code for each type of version
(definition and instance).
Only attempt migrations if there is a version difference.
Remove the function to get the add-on version as it is no longer needed.