Disable active reflection probe operators when linked.
This is more consistent with how Blender does things and
saves the user a click, while still displaying to the user
the reason why the operator is disabled via a tooltip.

Note: only operations relating to the active reflection
probe are disabled.  Operations that work on multiple probes
are still enabled and will warn if there are no valid probes
to operate on when executed.