Stabilize the report viewer.
Keep the report viewer window a consistent size and evenly pad the space
between messages, except when there are less than 4 messages in which case
all extra padding is added to the bottom and only a single line is used to
separate the messages.  If the messages can't be evenly spaced then
prioritize padding the bottom messages first.

15 lines are shown and a maximum of 5 messages.

Calculate/prepare the messages, and their padding, to be shown for each
starting index in advance when the report is first displayed and then
just display the messages that will fit in the window for the current
index.  If a message is larger than the allotted size, it will be the
only one displayed for that index and the window will stretch to
accommodate it.  Precalculating/preparing the messages keeps the whole
process fairly simple and is needed to accurately display the current
position in the report with the percentage bar since the number of
messages displayed for each index is now variable.

Improve variable names to prevent confusion.