Preserve property definition order when auto displaying/gathering.
The previous method for preventing internal HubsComponent base class
properties from sometimes being displayed in component subclasses didn't
preserve the order they were defined in and resulted in properties being
displayed differently each time in an unordered fashion.  This fixes it
so that the properties are displayed/gathered in the order they are defined.