Migrate components on add-on registration.
This allows components to be migrated if the add-on is enabled after
a file with Hubs components has already been loaded.
This migration is only triggered when the add-on is activated in the
middle of a session.  An undo step is added to help prevent an
accidental loss of the migration, and a message is displayed to warn
the user if a migration actually occurs along with any local
migration messages.