Warn about and display components on unsupported hosts.
Display a warning message on migration about components on unsupported hosts.
Display a simple panel for the component, similar to what is done for unknown
components, and allow the user to remove the components.

A method is provided to allow custom messages for components on unsupported hosts.
Add a utility function to retrieve the host reference for warning messages that is
less hacky than the current method.