Fix poll errors during migration.
Refactor the component poll methods to receive
the component host, and optionally the object,
as an argument and use them directly, rather than
receiving the context, which isn't correct during
migrations, and using that to get the host/object.

Remove the warning message in the video-texture-source
component ui for no camera being found in the object
hierarchy.