Armature migration and gizmo fixes.
Migrate armatures separately from objects because armatures
can exist separately from objects.  Update bones from edit mode
so none are missed during migrations.  Warnings are shown for
armatures that aren't connected to objects, regardless of
migration type.
Refresh gizmos and account for gizmo errors when switching the
armature for an object.  Note: this doesn't refresh for all
situations, notably, undo/redo will still need a manual refresh
of gizmos.
Refactor the get_host_reference_message function to include
whether it's an object or just an armature that the bone is a part
of and refactor the video texture source component to use the
get_host_reference_message function.