Display forced export parameters and support Blender versions under 3.2.
Display the forced glTF export parameters with their values in the UI
of the scene debugger and include the reasoning on why they are forced
to that state in their tooltips.
Do not add `use_active_scene` to the export parameters or show it in
the UI for Blender versions below 3.2 because these versions don't
support it.