Switch to getting the node by the id rather than looking it up by name in Blender versions under 3.1.
This allows for nodes without a name to have their components imported and also makes sure the correct node is imported when there are more than one with the same name (Spoke can generate files with duplicate names).
Note: the old way of getting the node by name is left as a fallback, just in case.