Add Blender 4.2+ support for the import/export settings panels.
The glTF add-on changed how it handles drawing the user extension settings panels for the import/export dialogs.  Now the glTF add-on searches for a single draw method (which handles both import and export) in the __init__.py file of the Hubs add-on where previously it provided a panel for the importer and exporter respectively that we could hook into.  This commit adds support for the new method while preserving support for the previous method.