Fix the Navigation Mesh component export for Blender 4.2
Add an extra dummy value to the Navigation Mesh component when exporting from Blender 4.2.
Hubs expects the value for the Navigation Mesh to be an empty dictionary, but the glTF exporter attempts to strip these out.
To combat this, the Hubs exporter uses a dummy value that gets stripped down to an empty dictionary by the glTF exporter, but in Blender 4.2 the glTF exporter added in an additional call to strip out empty values and ended up removing the component entirely.

Fixes https://github.com/Hubs-Foundation/hubs-blender-exporter/issues/311