Fix setting up the camera for reflection probe baking in Blender 4.2
Blender has re-arranged it's Cycles API so that the panorama settings that were previously stored on the `cycles` property of the camera data are now stored directly on the camera data.  The root of the panorama settings is now abstracted into a variable so that the panorama settings can be accessed from Blender 4.2 while maintaining backward compatibility with previous versions.