Fix the default particle emitter source example URL and add a warning
The current default particle emitter source URL is an example URL that was supposed to resolve to the default particle if "example.org" was replaced with the domain of your instance; however the assets subdomain was erroneously omitted and it wasn't obvious that the default wouldn't work and what to change to make it work.
This commit adds the missing subdomain and a warning that the default URL won't resolve unmodified with an explanation on what to change to correctly point it to the default white square particle on your instance.